Fixed Debug builds, SetProperty errors, Fiddling
Fixed old library references for v8 compiled for Debug.
Fixed error handling for JavascriptExternal::SetProperty().
Removed array index setting regression test because it does not make
sense - we copy arrays given to v8 and so changes will not be reflected
in the C# original.
Added Fiddling project for easy debugging.