Made use in multiple AppDomains thread-safe
Unfortunately std::mutex is not allowed in CLR code. Using monitors on
static CLR Objects also won't do the trick since those will obviously
also differ for each AppDomain. Thus I ended up using a named mutex.

This is not ideal since it also locks over process boundaries (which is
not necessary in this case), but hopefully this will not hurt.

The atomic_flag is still required since otherwise a boolean flag might
live in a register, resulting in an outdated read.

Followup of
https://github.com/JavascriptNet/Javascript.Net/pull/53#issuecomment-416512050