Fixed date conversions
JavaScript usually does not care a lot about timezones, but since the
conversion happens via milliseconds since epoch (which is UTC based) one
needs to be a bit careful about which conversions to do where.

Added/moved tests to a central place so there is an overview over all
date related tests.