Add support for registered constructors
This allows to register a function (delegate in .NET) as constructor
function in JavaScript.
In order to support this a new method `SetConstructor` was introduced
which allows to register a constructor function for a specific type
under a name. This fuction is then available to the executed JavaScript
code and invoking `new` with this function results in the objects being
registered as `instanceof` the registered function.

Additionally, objects which are created in .NET code are also converted
to V8 while preserving the `instanceof` relation if the type is
registered.

Part of #60