Add support for the `Symbol.iterator` function on objects implementing `IEnumerable`
This adds support for the `for-of`-Loop and the spread operator on all
custom objects implementing `IEnumerable`. Objects not implementing
`IEnumerable` return `undefined`. The iterator function itself is cached
if the enumerable object stays the same; however, this has no impact if
the `IEnumerable` is always created new. The behavior of arrays and
`List<T>` is unchanged as they are converted to an array automatically.