Make sure zero bytes are not discarded when converting strings
Since C++ uses zero byte terminated strings, a cast to wchar_t* discards all characters after a zero byte when used directly to create a C# or V8 string. This is bad for strings containing a zero byte somewhere. We fix that by using a constructor overload taking the actual length of the string as parameter.