Create `JavascriptScope` when comparing two `JavascriptFunction` instances
If we don't do that `isolate->GetCurrentContext` returns `nullptr` resulting in a
crash. Because we are dealing with two distinct function instances we also assert
that they are from the same `JavascriptContext` and otherwise return `false` early.

The crash predominantly happens during high GC pressure. The comparison is
caused by the `Contains` check in `JavascriptExternal::GCCallback` during GC,
resulting in a `nullptr` context in the equality operator of `JavascriptFunction`
because of the missing scope.